moved all filter things to convster
